# perf(tspath): optimize `hasRelativePathSegment`