This pull request optimises `hasRelativePathSegment`, delivering a ~56% perf improvement.

This function is Called `27_628_418` times via tsgolint when linting vscode repo.

```
                                                     │   old.txt   │               new.txt               │
                                                     │   sec/op    │   sec/op     vs base                │
HasRelativePathSegment/foo/bar/baz-12                  29.10n ± 2%   14.79n ± 1%  -49.16% (p=0.000 n=10)
HasRelativePathSegment/./some/path-12                  2.065n ± 0%   2.039n ± 2%   -1.26% (p=0.022 n=10)
HasRelativePathSegment//foo/./bar/../../.-12           7.030n ± 0%   2.025n ± 0%  -71.20% (p=0.000 n=10)
HasRelativePathSegment/foo/foo/foo/foo/foo/...etc-12   9.076n ± 1%   2.308n ± 1%  -74.56% (p=0.000 n=10)
geomean                                                7.869n        3.446n       -56.21%
```

this function has tests here to ensure we don't regress vs the old version:

https://github.com/microsoft/typescript-go/blob/d6c086714733161b1c36448ba1fc5e1b72ff233b/internal/tspath/path_test.go#L610-L618